8 - Collect preset names via SysEx
Got the correct reset ID going out to the module, and processing the incoming values correctly too.
